chatbot_backend/requirements.txt
fastapi
uvicorn
openai
pymavlink
python-multipart
python-dotenv
